applied PR feedback II
