+++ frontend/javascripts/libs/utils.ts
@@ -14,16 +14,7 @@ import type {
   TypedArray,
 } from "oxalis/constants";
 import window, { document, location } from "libs/window";
-import { ArbitraryObject } from "types/globals";
-
-export type Comparator<T> = (arg0: T, arg1: T) => -1 | 0 | 1;
-export type ArrayElement<A> = A extends readonly (infer T)[] ? T : never;
-export type NestedOmit<T, K extends PropertyKey> = {
-  [P in keyof T as P extends K ? never : P]: NestedOmit<
-    T[P],
-    K extends `${Exclude<P, symbol>}.${infer R}` ? R : never
-  >;
-};
+import { ArbitraryObject, Comparator } from "types/globals";
 
 type UrlParams = Record<string, string>;
 // Fix JS modulo bug
