+++ frontend/javascripts/types/globals.d.ts
@@ -5,8 +5,15 @@ declare global {
 }
 
 // https://stackoverflow.com/questions/49285864/is-there-a-valueof-similar-to-keyof-in-typescript
-type ValueOf<T> = T[keyof T];
-type EmptyObject = Record<string, never>;
-type ArbitraryObject = Record<string, any>;
-type ArbitraryFunction = (...args: Array<any>) => any;
-export { ValueOf, EmptyObject, ArbitraryObject, ArbitraryFunction };
+export type ValueOf<T> = T[keyof T];
+export type EmptyObject = Record<string, never>;
+export type ArbitraryObject = Record<string, any>;
+export type ArbitraryFunction = (...args: Array<any>) => any;
+export type Comparator<T> = (arg0: T, arg1: T) => -1 | 0 | 1;
+export type ArrayElement<A> = A extends readonly (infer T)[] ? T : never;
+export type NestedOmit<T, K extends PropertyKey> = {
+  [P in keyof T as P extends K ? never : P]: NestedOmit<
+    T[P],
+    K extends `${Exclude<P, symbol>}.${infer R}` ? R : never
+  >;
+};
